refactor(models): 正交分解 models 模块，提升单一职责与可维护性;
- plugin.py（264行/8类）拆分为 5 个正交子模块 + re-export 兼容层:
  plugin_common.py（枚举/权限）、mcp.py（Server/Tool 定义）、
  mcp_runtime.py（执行记录）、skill.py、sub_agent.py
- 从 pulse.py 提取 UserState/AppState 至 state.py（复合主键模式与 UUID 模式正交分离）
- knowledge_runtime.py 提取 _KnowledgeRunMixin 消除 KnowledgeGraphRun/KnowledgePipelineRun 的字段重复
- security.py 修复 DateTime(timezone=False) 为 TIMESTAMP 统一时间戳规范
- base.py 新增 DEFAULT_EMBEDDING_DIM 常量替换 4 处 Vector(1536) 魔法数字
- __init__.py 补全遗漏的 MemoryAutomationConfig/KnowledgeDocument/DEFAULT_EMBEDDING_DIM 导出
- 所有现有导入路径通过 re-export 兼容层保持向后兼容
