+++ apps/negentropy/src/negentropy/models/__init__.py
@@ -1,69 +1,70 @@
 from .action import SandboxExecution, Tool, ToolExecution
-from .base import NEGENTROPY_SCHEMA, Base, TimestampMixin, Vector, fk
-from .internalization import ConsolidationJob, Fact, Instruction, Memory, MemoryAuditLog
+from .base import DEFAULT_EMBEDDING_DIM, NEGENTROPY_SCHEMA, Base, TimestampMixin, Vector, fk
+from .internalization import ConsolidationJob, Fact, Instruction, Memory, MemoryAuditLog, MemoryAutomationConfig
 from .knowledge_runtime import KnowledgeGraphRun, KnowledgePipelineRun
-from .observability import Trace
-from .perception import Corpus, Knowledge
-from .plugin import (
-    McpServer,
-    McpTool,
-    McpToolRun,
-    McpToolRunEvent,
-    McpTrialAsset,
-    PluginPermission,
-    PluginPermissionType,
-    PluginVisibility,
-    Skill,
-    SubAgent,
-)
+from .mcp import McpServer, McpTool
+from .mcp_runtime import McpToolRun, McpToolRunEvent, McpTrialAsset
 from .model_config import ModelConfig, ModelType
-from .pulse import AppState, Event, Message, Run, Snapshot, Thread, UserState
+from .observability import Trace
+from .perception import Corpus, Knowledge, KnowledgeDocument
+from .plugin_common import PluginPermission, PluginPermissionType, PluginVisibility
+from .pulse import Event, Message, Run, Snapshot, Thread
 from .security import Credential
+from .skill import Skill
+from .state import AppState, UserState
+from .sub_agent import SubAgent
 
 __all__ = [
     # Base
     "Base",
+    "DEFAULT_EMBEDDING_DIM",
     "NEGENTROPY_SCHEMA",
     "TimestampMixin",
     "Vector",
     "fk",
-    # Pulse
+    # Pulse (会话)
     "Thread",
     "Event",
     "Run",
     "Message",
     "Snapshot",
+    # State (应用/用户状态)
     "UserState",
     "AppState",
-    # Internalization (was Hippocampus)
+    # Internalization (记忆)
     "Memory",
     "Fact",
     "ConsolidationJob",
     "Instruction",
     "MemoryAuditLog",
-    # Action (was Mind)
+    "MemoryAutomationConfig",
+    # Action
     "Tool",
     "ToolExecution",
     "SandboxExecution",
-    # Observability (was Mind)
+    # Observability
     "Trace",
-    # Perception
+    # Perception (知识)
     "Corpus",
     "Knowledge",
+    "KnowledgeDocument",
     # Knowledge Runtime
     "KnowledgeGraphRun",
     "KnowledgePipelineRun",
     # Security
     "Credential",
-    # Plugin
+    # Plugin Common
     "PluginVisibility",
     "PluginPermissionType",
     "PluginPermission",
+    # MCP
     "McpServer",
     "McpTool",
+    # MCP Runtime
     "McpToolRun",
     "McpToolRunEvent",
     "McpTrialAsset",
+    # Skill & SubAgent
     "Skill",
     "SubAgent",
     # Model Config
