+++ apps/negentropy/src/negentropy/models/knowledge_runtime.py
@@ -1,14 +1,20 @@
-from typing import Any, Dict, Optional
-from uuid import UUID
+"""知识运行时模型 -- Knowledge Graph 与 Pipeline 执行记录。"""
 
-from sqlalchemy import Integer, String, Text, UniqueConstraint
+from typing import Any, Dict, Optional
+
+from sqlalchemy import Integer, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
+
 from .base import NEGENTROPY_SCHEMA, Base, TimestampMixin, UUIDMixin
 
 
-class KnowledgeGraphRun(Base, UUIDMixin, TimestampMixin):
-    __tablename__ = "knowledge_graph_runs"
+class _KnowledgeRunMixin:
+    """知识运行记录共享字段。
+
+    提取 KnowledgeGraphRun 与 KnowledgePipelineRun 的公共列定义，
+    各子类仅需定义 __tablename__ 和 __table_args__。
+    """
 
     app_name: Mapped[str] = mapped_column(String(255), nullable=False)
     run_id: Mapped[str] = mapped_column(String(255), nullable=False)
@@ -16,6 +22,10 @@ class KnowledgeGraphRun(Base, UUIDMixin, TimestampMixin):
     payload: Mapped[Dict[str, Any]] = mapped_column(JSONB, nullable=False, server_default="{}")
     idempotency_key: Mapped[Optional[str]] = mapped_column(String(255))
     version: Mapped[int] = mapped_column(Integer, nullable=False, default=1, server_default="1")
+
+
+class KnowledgeGraphRun(_KnowledgeRunMixin, Base, UUIDMixin, TimestampMixin):
+    __tablename__ = "knowledge_graph_runs"
 
     __table_args__ = (
         UniqueConstraint("app_name", "run_id", name="knowledge_graph_runs_app_run_unique"),
@@ -24,23 +34,11 @@ class KnowledgeGraphRun(Base, UUIDMixin, TimestampMixin):
     )
 
 
-class KnowledgePipelineRun(Base, UUIDMixin, TimestampMixin):
+class KnowledgePipelineRun(_KnowledgeRunMixin, Base, UUIDMixin, TimestampMixin):
     __tablename__ = "knowledge_pipeline_runs"
-
-    app_name: Mapped[str] = mapped_column(String(255), nullable=False)
-    run_id: Mapped[str] = mapped_column(String(255), nullable=False)
-    status: Mapped[str] = mapped_column(String(20), nullable=False, default="pending", server_default="'pending'")
-    payload: Mapped[Dict[str, Any]] = mapped_column(JSONB, nullable=False, server_default="{}")
-    idempotency_key: Mapped[Optional[str]] = mapped_column(String(255))
-    version: Mapped[int] = mapped_column(Integer, nullable=False, default=1, server_default="1")
 
     __table_args__ = (
         UniqueConstraint("app_name", "run_id", name="knowledge_pipeline_runs_app_run_unique"),
         UniqueConstraint("app_name", "idempotency_key", name="knowledge_pipeline_runs_idempotency_unique"),
         {"schema": NEGENTROPY_SCHEMA},
     )
-
-
-# MemoryAuditLog 已迁移至 models/internalization.py
-# 参考: https://github.com/ThreeFish-AI/agentic-ai-cognizes/blob/master/docs/research/034-knowledge-base.md
-# Memory Governance 属于 Memory 模块，而非 Knowledge 模块
